% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_cheking_conversion.R
\name{error_checking_conversion}
\alias{error_checking_conversion}
\title{Error Checking}
\usage{
error_checking_conversion(dataset)
}
\arguments{
\item{dataset}{The dataset to be checked.}
}
\value{
If `dataset` is a list of 2-vectors, a dataframe with two columns (with numerical entries), or a vector with an even number of numeric entries, then a list of 2-vectors is returned. Otherwise an error message is returned.
}
\description{
This function converts vectors and dataframes into lists, and checks for any errors.
}
