% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_hull.R
\name{in_hull}
\alias{in_hull}
\title{Determines if Points are in the Convex Hull}
\usage{
in_hull(dataset, pointset)
}
\arguments{
\item{dataset}{The main set of points - a list of vectors of length 2, each representing a point in 2 dimensions. May also accept a dataframe or vector.}

\item{pointset}{The points to check if they are in the convex hull - a list of vectors of length 2, each representing a point in 2 dimensions. May also accept a dataframe or vector.}
}
\value{
A list where each entry corresponds to the respective entry in the pointset, stating the determination for that point. The determination is either "a vertex of the convex hull", "on an edge of the convex hull (but not a vertex)", "inside the convex hull, or "outside the convex hull".
}
\description{
This function checks if points in a list are in the convex hull of a set of points.
}
