% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarvis_march.R
\name{jarvis_march}
\alias{jarvis_march}
\title{Runs the Jarvis-March Algorithm}
\usage{
jarvis_march(dataset)
}
\arguments{
\item{dataset}{A list of vectors of length 2, each representing a point in 2 dimensions. May also accept a dataframe or vector.}
}
\value{
A list of vectors of length 4, each representing a line segment of the convex hull.
}
\description{
This function runs the Jarvis-March algorithm on a set of 2-dimensional points.
}
